Hive sprint 3 web choix utilisateur

+++ src/pages/Association/Profile/ModifyProfile.tsx
@@ -1,11 +1,9 @@
-import "./ModifyProfile.scss";
 import React, { useState, useEffect } from "react";
 import { useNavigate } from "react-router-dom";
 import { Container, Row, Col } from "react-bootstrap";
-import { Password } from "@mui/icons-material";
-import config from '../../../config';
+import './ModifyProfile.scss';
 
-const src_img = require('../../../images/titleLogo.png');
+const src_img = require('../../../images/logo/primary_logo.png');
 
 type newProfile = {
     name: string,
@@ -28,7 +26,7 @@ function ModifyProfilePage(props: any) {
   useEffect(() => {
     console.log(localStorage)
     const getProfile = () => {
-      let url = `${config.apiUrl}/associations/profile`;
+      let url = 'http://localhost:8000/associations/profile';
       fetch(url, {
         method: 'GET',
         headers: {
@@ -63,11 +61,11 @@ function ModifyProfilePage(props: any) {
     const emailRegex = /^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}$/;
     return emailRegex.test(email);
 }
-
+/* 
   function validatePhone(phone: string): boolean {
     const re = /^(?:(?:\+|00)33[\s.-]{0,3}(?:\(0\)[\s.-]{0,3})?|0)[1-9](?:(?:[\s.-]?\d{2}){4}|\d{2}(?:[\s.-]?\d{3}){2})$/;
     return re.test(phone);
-  }
+  } */
 
   const handleFileChange = (event: React.ChangeEvent<HTMLInputElement>) => {
      const file = event.target.files ? event.target.files[0] : null;
@@ -79,7 +77,8 @@ function ModifyProfilePage(props: any) {
          setProfilePicture(dataUrl);
          const formData = new FormData();
          formData.append('file', file);
-         fetch(`${config.apiUrl}/associations/profile/`, {
+         const url = 'http://localhost:8000/associations/profile/';
+         fetch(url, {
            method: 'POST',
            headers: {
              'Content-Type': 'application/json',
@@ -119,7 +118,8 @@ function ModifyProfilePage(props: any) {
         profile_picture: profile_picture,
     }
 
-    fetch(`${config.apiUrl}/associations/update`, {
+    let url = 'http://localhost:8000/associations/update';
+    fetch(url, {
       method: 'POST',
       headers: {
         'Content-Type': 'application/json',
@@ -136,9 +136,9 @@ function ModifyProfilePage(props: any) {
 
   /* Function to add when back is gonna be done */
 
-  const deleteAccount = () => {
-    /* if (window.confirm('Are you sure you want to delete your account?')) {
-      let url = `${config.apiUrl}/associations/profile`;
+  /*const deleteAccount = () => {
+     if (window.confirm('Are you sure you want to delete your account?')) {
+      let url = 'http://localhost:8000/associations/profile';
       fetch(url, {
         method: 'DELETE',
         headers: {
@@ -157,8 +157,8 @@ function ModifyProfilePage(props: any) {
         .catch((error) => 
           console.log(error);
         });
-    } */
-  };
+    }
+  };*/
 
   return (
     <Container className="profile-container">
